+++ packages/demo-cra/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.17](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.17-alpha.1...demo-cra@1.0.17) (2020-02-18)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.17-alpha.1](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.17-alpha.0...demo-cra@1.0.17-alpha.1) (2020-02-16)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.17-alpha.0](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.16...demo-cra@1.0.17-alpha.0) (2020-02-14)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
 ## [1.0.16](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.16-alpha.1...demo-cra@1.0.16) (2020-02-11)
 
 **Note:** Version bump only for package demo-cra