+++ packages/demo-next/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.2.3-alpha.1...nextjs-blog-starter@0.2.3) (2020-02-18)
+
+**Note:** Version bump only for package nextjs-blog-starter
+
+
+
+
+
+## [0.2.3-alpha.1](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.2.3-alpha.0...nextjs-blog-starter@0.2.3-alpha.1) (2020-02-16)
+
+**Note:** Version bump only for package nextjs-blog-starter
+
+
+
+
+
+## [0.2.3-alpha.0](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.2.2...nextjs-blog-starter@0.2.3-alpha.0) (2020-02-14)
+
+**Note:** Version bump only for package nextjs-blog-starter
+
+
+
+
+
 ## [0.2.2](https://github.com/kendallstrautman/nextjs-blog-starter/compare/nextjs-blog-starter@0.2.2-alpha.1...nextjs-blog-starter@0.2.2) (2020-02-11)
 
 **Note:** Version bump only for package nextjs-blog-starter