+++ packages/gatsby-plugin-tinacms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.24](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.24-alpha.1...gatsby-plugin-tinacms@0.1.24) (2020-02-18)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
+## [0.1.24-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.24-alpha.0...gatsby-plugin-tinacms@0.1.24-alpha.1) (2020-02-16)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
+## [0.1.24-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.23...gatsby-plugin-tinacms@0.1.24-alpha.0) (2020-02-14)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
 ## [0.1.23](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.23-alpha.1...gatsby-plugin-tinacms@0.1.23) (2020-02-11)
 
 **Note:** Version bump only for package gatsby-plugin-tinacms