+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.2-alpha.1...next-tinacms-markdown@0.2.2) (2020-02-18)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.2-alpha.1](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.2-alpha.0...next-tinacms-markdown@0.2.2-alpha.1) (2020-02-16)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.2-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.1...next-tinacms-markdown@0.2.2-alpha.0) (2020-02-14)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
 ## [0.2.1](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.1-alpha.1...next-tinacms-markdown@0.2.1) (2020-02-11)
 
 **Note:** Version bump only for package next-tinacms-markdown